+++ app/Http/Controllers/api/v1/RoomFileController.php
@@ -8,10 +8,12 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\RoomFileIndexRequest;
 use App\Http\Requests\StoreRoomFileRequest;
 use App\Http\Requests\UpdateRoomFileRequest;
+use App\Http\Requests\UpdateRoomSystemDefaultPresentation;
 use App\Http\Resources\PrivateRoomFileResource;
 use App\Http\Resources\RoomFileResource;
 use App\Models\Room;
 use App\Models\RoomFile;
+use App\Settings\BigBlueButtonSettings;
 use App\Settings\GeneralSettings;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
@@ -72,6 +74,11 @@ class RoomFileController extends Controller
         // If user is allowed to view all files, return PrivateRoomFile resource to show additional information
         if (Gate::allows('viewAllFiles', $room)) {
             $additional['default'] = $room->files()->where('default', true)->first();
+            $additional['system_default'] = [
+                'file' => app(BigBlueButtonSettings::class)->default_presentation,
+                'use_in_meeting' => $room->use_system_default_presentation_in_meeting,
+                'use_as_default' => $room->use_system_default_presentation_as_default,
+            ];
 
             return PrivateRoomFileResource::collection($resource->paginate(app(GeneralSettings::class)->pagination_page_size))->additional($additional);
         }
@@ -108,27 +115,58 @@ class RoomFileController extends Controller
     public function update(UpdateRoomFileRequest $request, Room $room, RoomFile $file)
     {
         if ($request->has('use_in_meeting')) {
-            $file->use_in_meeting = $request->use_in_meeting;
-            // If no default file for this room is set, set this file as default
-            if (! $room->files()->where('default', true)->exists()) {
-                $file->default = true;
-            }
+            $file->use_in_meeting = $request->boolean('use_in_meeting');
         }
 
         if ($request->has('download')) {
-            $file->download = $request->download;
+            $file->download = $request->boolean('download');
         }
 
-        if ($request->has('default') && $request->default === true) {
-            // Make other files not the default
-            $room->files()->update(['default' => false]);
-            // Set this file as default
-            $file->default = true;
+        if ($request->has('default')) {
+            if ($request->boolean('default') === false) {
+                $file->default = false;
+            } else {
+                // Make other files not the default
+                $room->files()->whereNot('id', $file->id)->update(['default' => false]);
+                // Set this file as default
+                $file->default = true;
+
+                // If a file is set as default, the system default presentation must not be used as default
+                $room->use_system_default_presentation_as_default = false;
+                $room->save();
+            }
         }
 
         $file->save();
 
         Log::info('Changed file settings for file {file} in room {room}', ['room' => $room->getLogLabel(), 'file' => $file->getLogLabel()]);
+
+        $room->updateDefaultFile();
+
+        return response()->noContent();
+    }
+
+    public function updateSystemDefault(UpdateRoomSystemDefaultPresentation $request, Room $room)
+    {
+        if ($request->has('use_in_meeting')) {
+            $room->use_system_default_presentation_in_meeting = $request->use_in_meeting;
+        }
+
+        if ($request->has('default')) {
+            // Make other files not the default
+            if ($request->default === true) {
+                $room->files()->update(['default' => false]);
+
+                // If system default is set as default, it must also be used in the next meeting
+                $room->use_system_default_presentation_in_meeting = true;
+            }
+
+            $room->use_system_default_presentation_as_default = $request->default;
+        }
+
+        $room->save();
+
+        Log::info('Changed system default presentation settings in room {room}', ['room' => $room->getLogLabel()]);
 
         $room->updateDefaultFile();