+++ resources/js/components/RoomTabFilesConfigureButton.vue
@@ -1,12 +1,12 @@
 <template>
   <!-- button -->
   <Button
-    v-tooltip="$t('rooms.files.edit')"
-    :aria-label="$t('rooms.files.edit_aria', { filename: filename })"
+    v-tooltip="$t('rooms.files.configure')"
+    :aria-label="$t('rooms.files.configure_aria', { filename: filename })"
     :disabled="disabled"
     severity="info"
     icon="fa-solid fa-edit"
-    data-test="room-files-edit-button"
+    data-test="room-files-configure-button"
     @click="showModal"
   />
 
@@ -14,14 +14,14 @@
   <Dialog
     v-model:visible="modalVisible"
     modal
-    :header="$t('rooms.files.edit')"
+    :header="$t('rooms.files.configure_dialog_title', { name: filename })"
     :style="{ width: '500px' }"
     :breakpoints="{ '575px': '90vw' }"
     :draggable="false"
     :close-on-escape="!isLoadingAction"
     :dismissable-mask="false"
     :closable="!isLoadingAction"
-    data-test="room-files-edit-dialog"
+    data-test="room-files-configure-dialog"
   >
     <template #footer>
       <div class="flex justify-end gap-2">
@@ -37,39 +37,27 @@
           severity="success"
           :loading="isLoadingAction"
           data-test="dialog-save-button"
-          form="room-files-edit-form"
+          form="room-files-configure-form"
           type="submit"
         />
       </div>
     </template>
 
-    <Form id="room-files-edit-form" :disabled="isLoadingAction" @submit="save">
-      <div class="field grid grid-cols-12 gap-4" data-test="download-field">
-        <label for="download" class="col-span-12 mb-2 md:col-span-6 md:mb-0">
-          {{ $t("rooms.files.downloadable") }}
-        </label>
-        <div class="col-span-12 md:col-span-6">
-          <ToggleSwitch
-            v-model="newDownload"
-            input-id="download"
-            required
-            :disabled="isLoadingAction"
-            :invalid="formErrors.fieldInvalid('download')"
-          />
-          <FormError :errors="formErrors.fieldError('download')" />
-        </div>
-      </div>
-
+    <Form
+      id="room-files-configure-form"
+      :disabled="isLoadingAction"
+      @submit="save"
+    >
       <div
-        class="field grid grid-cols-12 gap-4"
+        class="field mt-2 grid grid-cols-12 gap-4"
         data-test="use-in-meeting-field"
       >
         <label
           for="use_in_meeting"
-          class="col-span-12 mb-2 md:col-span-6 md:mb-0"
-          >{{ $t("rooms.files.use_in_next_meeting") }}</label
+          class="col-span-12 mb-2 md:col-span-8 md:mb-0"
+          >{{ $t("rooms.files.available_in_next_meeting") }}</label
         >
-        <div class="col-span-12 md:col-span-6">
+        <div class="col-span-12 md:col-span-4">
           <ToggleSwitch
             v-model="newUseInMeeting"
             input-id="use_in_meeting"
@@ -81,11 +69,11 @@
         </div>
       </div>
 
-      <div class="field grid grid-cols-12 gap-4" data-test="default-field">
-        <label for="default" class="col-span-12 mb-2 md:col-span-6 md:mb-0">
-          {{ $t("rooms.files.default") }}
-        </label>
-        <div class="col-span-12 md:col-span-6">
+      <div class="field mt-2 grid grid-cols-12 gap-4" data-test="default-field">
+        <label for="default" class="col-span-12 mb-2 md:col-span-8 md:mb-0">{{
+          $t("rooms.files.default")
+        }}</label>
+        <div class="col-span-12 md:col-span-4">
           <ToggleSwitch
             v-model="newDefault"
             input-id="default"
@@ -94,6 +82,25 @@
             :invalid="formErrors.fieldInvalid('default')"
           />
           <FormError :errors="formErrors.fieldError('default')" />
+        </div>
+      </div>
+
+      <div
+        class="field mt-2 grid grid-cols-12 gap-4"
+        data-test="download-field"
+      >
+        <label for="download" class="col-span-12 mb-2 md:col-span-8 md:mb-0">{{
+          $t("rooms.files.downloadable")
+        }}</label>
+        <div class="col-span-12 md:col-span-4">
+          <ToggleSwitch
+            v-model="newDownload"
+            input-id="download"
+            required
+            :disabled="isLoadingAction"
+            :invalid="formErrors.fieldInvalid('download')"
+          />
+          <FormError :errors="formErrors.fieldError('download')" />
         </div>
       </div>
     </Form>
@@ -179,7 +186,7 @@ watch(newUseInMeeting, (value) => {
 });
 
 /**
- * Sends a request to the server to create a new token or edit a existing.
+ * Sends a request to the server to update the file configuration.
  */
 function save() {
   isLoadingAction.value = true;