+++ routes/api.php
@@ -138,6 +138,7 @@ Route::prefix('v1')->name('api.v1.')->group(function () {
             // File operations
             Route::middleware('can:manageFiles,room')->group(function () {
                 Route::post('rooms/{room}/files', [RoomFileController::class, 'store'])->name('rooms.files.add');
+                Route::put('rooms/{room}/files/system_default', [RoomFileController::class, 'updateSystemDefault'])->name('rooms.files.updateSystemDefault');
                 Route::put('rooms/{room}/files/{file}', [RoomFileController::class, 'update'])->name('rooms.files.update');
                 Route::delete('rooms/{room}/files/{file}', [RoomFileController::class, 'destroy'])->name('rooms.files.destroy');
             });